spark-server/requirements.txt
py4j==0.10.9.7
uvicorn==0.34.0
pydantic==2.10.6
fastapi==0.115.12
kafka-python==2.1.2
minio==7.1.0
numpy>=1.19.0  # Required for PySpark ML